ghosts, I blame ghosts (index was in the wrong package)
